return grocery list item on create

+++ groceryDB.py
@@ -268,10 +268,9 @@ class GroceryDB:
         rows = self.cursor.fetchall()
         return rows
     def updateGroceryListItem(self, listID, ingredientID, quantity, originalQuantityType, newQuantityType):
-        if self.groceryListItemExists(listID, ingredientID, newQuantityType):
+        if self.groceryListItemExists(listID, ingredientID, newQuantityType) and originalQuantityType != newQuantityType:
             existingItem = self.getGroceryListItem(listID, ingredientID, newQuantityType)
-            if originalQuantityType != newQuantityType:
-                self.deleteGroceryListItem(listID, ingredientID, newQuantityType) # delete the previously existing item
+            self.deleteGroceryListItem(listID, ingredientID, newQuantityType) # delete the previously existing item
             newQuantity = addQuantityStrings(quantity, existingItem["quantity"])
             queryString = "UPDATE grocery_list_items SET quantity = %s, quantity_type = %s WHERE list_id = %s AND ingredient_id = %s AND quantity_type = %s"
             self.cursor.execute(queryString, (newQuantity, newQuantityType, listID, ingredientID, originalQuantityType))
